# backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Hackathon #TODO
  description: #TODO
  version: v1
  contact:
    name: Team Ctrl+Alt+Elite

components:
  schemas:
    ArticleDTO:
      description: DTO for main entity Article
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: article1
        author:
          type: string
          example: "0x35686896756"
        tags:
          type: array
          items:
            type: string
            example: "IT"
        text:
          type: string
          example: "Lorem impsum"
        # pathToFile: #TODO for pdf
      required:
        - title
        - tags
        - text

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentification

  parameters:
    Address:
      name: Address
      in: header
      required: true
      description: address in blockchain for auth
      schema:
        type: string
        example: "0x35686896756"

  responses:
    200Pong:
      description: pong
      content:
        application/json:
          schema:
            type: string
            example: "pong"

    400BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: string
            example: invalid body
    
    401Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: string
            example: error

security:
  - bearerAuth: []

tags:
  - name: Ping
    description: Tag for ping operation
  - name: auth
    description: Tag for authentification operations
  - name: Articles
    description: Tag for operations with articles

paths:
  /v1/ping:
    get:
      tags:
        - "Ping"
      summary: "Check if server is alive"
      description: "Returns pong as answer"
      operationId: pong
      responses:
        200:
          $ref: '#/components/responses/200Pong'

  /v1/auth/ping:
    get:
      tags:
        - "Ping"
      summary: "Ping with authentification"
      description: "Return pong if you are authentificated"
      operationId: authPong
      security:
        - bearerAuth: []
      responses:
        200:
          $ref: '#/components/responses/200Pong'
        401:
          $ref: '#/components/responses/401Unauthorized'

  /v1/auth/nonce:
    get:
      tags:
        - "Auth"
      summary: "Step of authentification"
      description: "Returns nonce for ethereum sign"
      operationId: authNonce
      parameters:
        - $ref: '#/components/parameters/Address'
      responses:
        200:
          description: message for signing
          content:
            application/json:
              schema:
                type: string
                example: nonce

  /v1/auth:
    post:
      tags:
        - "Auth"
      summary: "Step of authentification"
      description: "Return JWT, if nonce signed correctly"
      operationId: auth
      parameters:
        - $ref: '#/components/parameters/Address'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              example: SHA-256
      responses:
        200:
          description: successful auth
          content:
            application/json:
              schema:
                type: string
                example: "BABABABAB"
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'

  /v1/articles/new:
    post:
      tags:
        - "Articles"
      summary: "Creates new article"
      operationId: createArticle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleDTO'
      responses:
        201:
          description: article created!
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'

  /v1/articles/my:
    get:
      tags:
        - "Articles"
      summary: "Return yours articles"
      operationId: getArticles
      security:
        - bearerAuth: []
      responses:
        200:
          description: Articles which yoy authored
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleDTO'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'

  /v1/articles/search/{phrase}:
    get:
      tags:
        - "Articles"
      summary: "Search relevant Articles"
      operationId: searchArticles
      security:
        - bearerAuth: []
      parameters:
        - name: phrase
          in: path
          required: true
          schema:
            type: string
            example: "Merge sort"
      responses:
        200:
          description: Articles relevant to your phrase
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleDTO'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'

  /v1/articles/{id}:
    put:
      tags:
        - "Articles"
      summary: "Update chosen Article"
      operationId: updateArticle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "768ffrg8"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleDTO'
      responses:
        200:
          description: Article updated successefully
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          description: You are not author of this article
        404:
          description: Article with this id not found
    delete:
      tags:
        - "Articles"
      summary: "delete chosen Article"
      operationId: deleteArticle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "768ffrg8"
      responses:
        200:
          description: Article deleted successefully
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          description: You are not author of this article
        404:
          description: Article with this id not found



        
        